requirements.txt
Django==2.2
psycopg2-binary==2.7.7
uWSGI==2.0.17
djangorestframework==3.9.2
django-choices==1.6.2

factory-boy==2.11.1
pytest-django==3.4.8
django-cors-headers==2.5.2

celery==4.2.2
redis==3.2.1
